# Conflicts:
#	.idea/compiler.xml
#	Ev-System/pom.xml
#	Ev-System/src/main/java/com/example/Ev/System/controller/AppointmentController.java
#	Ev-System/src/main/java/com/example/Ev/System/entity/MaintenanceRecord.java
#	Ev-System/src/main/java/com/example/Ev/System/entity/Part.java
#	Ev-System/src/main/java/com/example/Ev/System/entity/PartUsage.java
#	Ev-System/src/main/java/com/example/Ev/System/repository/AppointmentRepository.java
#	Ev-System/src/main/java/com/example/Ev/System/repository/AppointmentServiceRepository.java
#	Ev-System/src/main/java/com/example/Ev/System/repository/ServiceCenterRepository.java
#	Ev-System/src/main/java/com/example/Ev/System/repository/ServiceTypeRepository.java
#	Ev-System/src/main/java/com/example/Ev/System/service/UserService.java
#	Ev-System/src/main/resources/application.properties

+++ Ev-System/src/main/java/com/example/Ev/System/repository/ServiceTypeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ServiceTypeRepository extends JpaRepository<ServiceType, Integer> {
+public interface ServiceTypeRepository extends JpaRepository<ServiceType, Long> {
     Optional<ServiceType> findById(Integer id);
 }
